refactor(workspaces): fixed minor comments
