+++ packages/server/modules/workspaces/tests/integration/tracking.spec.ts
@@ -22,7 +22,7 @@ import { db } from '@/db/knex'
 import { getPaginatedProjectModelsTotalCountFactory } from '@/modules/core/repositories/branches'
 import { updateAllWorkspacesTackingPropertiesFactory } from '@/modules/workspaces/services/tracking'
 import { logger } from '@/observability/logging'
-import { buildMixpanelFake } from '@/modules/shared/test/hepers/creation'
+import { buildMixpanelFake } from '@/modules/shared/test/helpers/mixpanel'
 import { expect } from 'chai'
 import { truncateTables } from '@/test/hooks'
 import { Workspaces } from '@/modules/workspaces/helpers/db'
